foodgram/requirements.txt
asgiref==3.6.0
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.4.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
gunicorn==20.0.4
psycopg2-binary==2.8.6
drf-extra-fields==3.4.0
python-dotenv==0.21.1